+++ src/deeplearning/softmax/std_softmax.cpp
@@ -5,13 +5,13 @@ namespace deeplearning {
 void StdSoftmax::Normalize(const std::vector<double> &input,
                            std::vector<double> &output) {
   long double sum = 0;
-  for (int i = 0; i < input.size(); i++) {
+  for (int i = 0; i < static_cast<int>(input.size()); i++) {
     sum += std::exp(input[i]);
   }
   if (sum == 0 || input.size() != output.size()) {
     return;
   }
-  for (int i = 0; i < input.size(); i++) {
+  for (int i = 0; i < static_cast<int>(input.size()); i++) {
     output[i] = std::exp(input[i]) / sum;
   }
 }
